refactor: break out websocket request handler to separate file.
